# Optimizes the incremental reader's symbol table reading logic for tables with imports.