… src/com/amazon/ion/impl/IonReaderBinaryIncremental.java