*Description of changes:*

Profiling has revealed that the incremental reader's existing handling of shared symbol tables performs poorly, and becomes noticeable when reading small values with shared symbol tables or streams that contain many symbol tables with imports.

This inefficiency was due to the fact that the reader stored symbols in a flat list, which required all imported symbols to be copied into the list every time the symbol table changed. After this change, only the local symbols are stored in that list; imported symbols are referenced from their underlying shared SymbolTables via [LocalSymbolTableImports](https://github.com/amzn/ion-java/blob/master/src/com/amazon/ion/impl/LocalSymbolTableImports.java), which is also used for this purpose by the other binary IonReader implementation.

I used ion-java-benchmark-cli to verify that there is no observed impact to performance for small or large streams that don't use shared symbol tables.

I also used ion-java-benchmark-cli to do a comparison targeted to this fix. It involves reading a small stream using shared symbol tables 100 times in a row.

Before:

```
Benchmark                                      (input)                                                                               (options)  Mode  Cnt       Score      Error   Units
Bench.run                                   shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4     151.307 ±    2.554   us/op
Bench.run:Heap usage                        shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4      94.836 ±  343.822      MB
Bench.run:Serialized size                   shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4      ≈ 10⁻³                 MB
Bench.run:·gc.alloc.rate                    shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4    4237.355 ±   69.259  MB/sec
Bench.run:·gc.alloc.rate.norm               shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4  706416.008 ±    0.022    B/op
Bench.run:·gc.churn.PS_Eden_Space           shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4    4230.845 ±   69.936  MB/sec
Bench.run:·gc.churn.PS_Eden_Space.norm      shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4  705330.862 ± 4978.273    B/op
Bench.run:·gc.churn.PS_Survivor_Space       shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4       0.203 ±    0.173  MB/sec
Bench.run:·gc.churn.PS_Survivor_Space.norm  shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4      33.794 ±   29.305    B/op
Bench.run:·gc.count                         shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4     883.000             counts
Bench.run:·gc.time                          shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4     355.000                 ms
```

After:

```
Benchmark                                      (input)                                                                               (options)  Mode  Cnt       Score       Error   Units
Bench.run                                   shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4     115.972 ±     2.762   us/op
Bench.run:Heap usage                        shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4     113.331 ±   485.528      MB
Bench.run:Serialized size                   shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4      ≈ 10⁻³                  MB
Bench.run:·gc.alloc.rate                    shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4    5253.101 ±   129.797  MB/sec
Bench.run:·gc.alloc.rate.norm               shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4  671216.006 ±     0.017    B/op
Bench.run:·gc.churn.PS_Eden_Space           shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4    5245.017 ±    95.518  MB/sec
Bench.run:·gc.churn.PS_Eden_Space.norm      shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4  670185.145 ±  5759.997    B/op
Bench.run:·gc.churn.PS_Survivor_Space       shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4       0.192 ±     0.184  MB/sec
Bench.run:·gc.churn.PS_Survivor_Space.norm  shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4      24.501 ±    23.765    B/op
Bench.run:·gc.count                         shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4     885.000              counts
Bench.run:·gc.time                          shape2.ion   read::{f:ION_BINARY,t:BUFFER,c:"imports.ion",a:STREAMING,k:true,R:INCREMENTAL,Z:4096}  avgt    4     361.000                  ms
```

The improvement was 151us -> 116us (23%).

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.